pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "aero-prop-logic-harness"
version = "0.1.0"
description = "AeroProp Logic Harness — engineering knowledge harness for civil aviation propulsion system control logic"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "Proprietary — Internal Engineering Use"}
authors = [
    {name = "Propulsion Control Systems Engineering Team"}
]

dependencies = [
    "pydantic>=2.5,<3",
    "ruamel.yaml>=0.18,<1",
    "typer>=0.9,<1",
    "rich>=13.0,<14",
    "jsonschema>=4.20,<5",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
]

[project.scripts]
aplh = "aero_prop_logic_harness.cli:app"

[tool.setuptools.packages.find]
include = ["aero_prop_logic_harness*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
